create branch 4.2

+++ src/usr.bin/systat/extern.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: extern.h,v 1.17 2007/02/24 09:45:27 otto Exp $	*/
+/*	$OpenBSD: extern.h,v 1.19 2007/08/09 02:38:09 ray Exp $	*/
 /*	$NetBSD: extern.h,v 1.3 1996/05/10 23:16:34 thorpej Exp $	*/
 
 /*-
@@ -48,9 +48,9 @@ extern int	*dk_select;
 extern int	CMDLINE;
 extern int	dk_ndrive;
 extern int	hz, stathz;
-extern u_int	naptime;
-extern int	nhosts;
-extern int	nports;
+extern double	naptime;
+extern size_t	nhosts;
+extern size_t	nports;
 extern int	protos;
 extern int	verbose;
 extern int	nflag;
